[#690] Move Custom**Marker into main builder namespace
