+++ iceoryx2/src/pending_response.rs
@@ -66,7 +66,7 @@ use crate::port::details::chunk::Chunk;
 use crate::port::details::chunk_details::ChunkDetails;
 use crate::port::update_connections::ConnectionFailure;
 use crate::raw_sample::RawSample;
-use crate::service::builder::publish_subscribe::CustomPayloadMarker;
+use crate::service::builder::CustomPayloadMarker;
 use crate::{port::ReceiveError, request_mut::RequestMut, response::Response, service};
 
 /// Represents an active connection to all [`Server`](crate::port::server::Server)
