+++ iceoryx2/src/service/header/request_response.rs
@@ -33,6 +33,15 @@ impl RequestHeader {
         self.client_port_id
     }
 
+    /// Returns how many elements are stored inside the requests's payload.
+    ///
+    /// # Details when using
+    /// [`CustomPayloadMarker`](crate::service::builder::publish_subscribe::CustomPayloadMarker)
+    ///
+    /// In this case the number of elements relates to the element defined in the
+    /// [`MessageTypeDetails`](crate::service::static_config::message_type_details::MessageTypeDetails).
+    /// When the element has a `payload.size == 40` and the `RequestMut::payload().len() == 120` it
+    /// means that it contains 3 elements (3 * 40 == 120).
     pub fn number_of_elements(&self) -> u64 {
         self.number_of_elements
     }
@@ -55,6 +64,15 @@ impl ResponseHeader {
         self.server_port_id
     }
 
+    /// Returns how many elements are stored inside the response's payload.
+    ///
+    /// # Details when using
+    /// [`CustomPayloadMarker`](crate::service::builder::publish_subscribe::CustomPayloadMarker)
+    ///
+    /// In this case the number of elements relates to the element defined in the
+    /// [`MessageTypeDetails`](crate::service::static_config::message_type_details::MessageTypeDetails).
+    /// When the element has a `payload.size == 40` and the `ResponseMut::payload().len() == 120` it
+    /// means that it contains 3 elements (3 * 40 == 120).
     pub fn number_of_elements(&self) -> u64 {
         self.number_of_elements
     }
