+++ iceoryx2/tests/publisher_tests.rs
@@ -19,7 +19,7 @@ mod publisher {
 
     use iceoryx2::port::{publisher::PublisherCreateError, LoanError};
     use iceoryx2::prelude::*;
-    use iceoryx2::service::builder::publish_subscribe::CustomPayloadMarker;
+    use iceoryx2::service::builder::CustomPayloadMarker;
     use iceoryx2::service::static_config::message_type_details::{TypeDetail, TypeVariant};
     use iceoryx2::service::{service_name::ServiceName, Service};
     use iceoryx2::testing::*;
