+++ iceoryx2/tests/service_request_response_tests.rs
@@ -19,7 +19,7 @@ mod service_request_response {
     use iceoryx2::port::server::Server;
     use iceoryx2::port::LoanError;
     use iceoryx2::prelude::{PortFactory, *};
-    use iceoryx2::service::builder::publish_subscribe::{CustomHeaderMarker, CustomPayloadMarker};
+    use iceoryx2::service::builder::{CustomHeaderMarker, CustomPayloadMarker};
     use iceoryx2::service::static_config::message_type_details::{TypeDetail, TypeVariant};
     use iceoryx2::testing::*;
     use iceoryx2_bb_testing::assert_that;
